src/main/java/de/nava/mlsample/domain/Products.java
package de.nava.mlsample.domain;

import javax.xml.bind.annotation.XmlAccessType;
import javax.xml.bind.annotation.XmlAccessorType;
import javax.xml.bind.annotation.XmlElement;
import javax.xml.bind.annotation.XmlRootElement;
import java.util.Collection;

@XmlRootElement(name = "products")
@XmlAccessorType(XmlAccessType.NONE)
public class Products {

    private Collection<Product> products;

    @XmlElement(name = "product")
    public Collection<Product> getProducts() {
        return products;
    }

    public void setProducts(Collection<Product> products) {
        this.products = products;
    }
}
